Refactor LessonCard to separate component and improve readability
- Extract LessonCard component to app/src/components/LessonCard.tsx
- Increase font sizes (description: text-base→text-lg, labels: text-xs→text-sm)
- Center align card text
- Change word count display from "ワード" to "全N語"
