+++ app/src/pages/Home.tsx
@@ -1,75 +1,11 @@
 import { useState } from 'react'
-import { Link } from 'react-router-dom'
 import lessonData from '../data/lesson.json'
 import { getBestScore } from '../lib/scores'
+import LessonCard from '../components/LessonCard'
 
 const lessons = Object.entries(lessonData)
 
 type Tab = 'challenge' | 'perfect'
-
-const CARD_COLORS = [
-  'from-pink-400 to-rose-500',
-  'from-orange-400 to-amber-500',
-  'from-yellow-400 to-orange-400',
-  'from-green-400 to-emerald-500',
-  'from-teal-400 to-cyan-500',
-  'from-blue-400 to-indigo-500',
-  'from-purple-400 to-violet-500',
-  'from-fuchsia-400 to-pink-500',
-]
-
-function getBadge(best: ReturnType<typeof getBestScore>) {
-  if (!best) return null
-  if (best.accuracy === 100) return { text: 'パーフェクト！', icon: '\uD83D\uDC51', bg: 'bg-yellow-400 text-yellow-900' }
-  if (best.accuracy >= 90) return { text: `${best.accuracy}%`, icon: '\u2B50', bg: 'bg-white/90 text-yellow-500' }
-  if (best.accuracy >= 70) return { text: `${best.accuracy}%`, icon: '\u25CB', bg: 'bg-white/90 text-blue-500' }
-  return { text: `${best.accuracy}%`, icon: '\u25B3', bg: 'bg-white/90 text-gray-600' }
-}
-
-function LessonCard({ id, lesson, index }: { id: string; lesson: (typeof lessonData)[keyof typeof lessonData]; index: number }) {
-  const best = getBestScore(id)
-  const badge = getBadge(best)
-
-  return (
-    <Link
-      key={id}
-      to={`/lesson/${id}`}
-      className={`group relative bg-linear-to-br ${CARD_COLORS[index % CARD_COLORS.length]} rounded-2xl p-4 text-white shadow-lg hover:shadow-xl transform hover:scale-105 transition-all duration-200 overflow-hidden`}
-    >
-      {badge && (
-        <div className={`absolute top-2 right-2 ${badge.bg} rounded-full px-2.5 py-1 flex items-center gap-1 text-xs font-bold shadow`}>
-          <span>{badge.icon}</span>
-          <span>{badge.text}</span>
-        </div>
-      )}
-
-      <div className="text-4xl mb-2 drop-shadow-md group-hover:animate-bounce">
-        {lesson.emoji ?? '\uD83D\uDCD6'}
-      </div>
-      <div className="text-xs font-medium opacity-80">
-        Lesson {index}
-      </div>
-      <div className="text-base font-bold mt-1 leading-tight">
-        {lesson.description}
-      </div>
-      <div className="text-xs opacity-70 mt-1">
-        {lesson.words.length} ワード
-      </div>
-
-      {!best && (
-        <div className="mt-2">
-          <span className="text-xs bg-white text-gray-700 font-bold px-3 py-1 rounded-full shadow-sm animate-pulse">
-            → チャレンジ！
-          </span>
-        </div>
-      )}
-
-      <div className="absolute -right-3 -bottom-3 text-7xl opacity-10 group-hover:opacity-20 transition-opacity">
-        {lesson.emoji ?? '\uD83D\uDCD6'}
-      </div>
-    </Link>
-  )
-}
 
 export default function Home() {
   const [tab, setTab] = useState<Tab>('challenge')
@@ -117,7 +53,7 @@ export default function Home() {
               : 'bg-gray-100 text-gray-500 hover:bg-gray-200'
           }`}
         >
-          {'\uD83D\uDC51'} パーフェクト
+          👑 パーフェクト
           {perfectCount > 0 && (
             <span className="ml-1.5 bg-white/30 px-1.5 py-0.5 rounded-full text-xs">
               {perfectCount}
